+++ packages/visual-stack-docs/src/containers/Layouts/DialogLayout/index.js
@@ -18,7 +18,6 @@ import {
   TextField,
 } from '@cjdev/visual-stack/lib/components/Form';
 import { Demo, MarginTopSnippet as Snippet } from '../../../components/Demo';
-import CJLogo from '@cjdev/visual-stack/lib/components/CJLogo';
 import './index.css';
 import {
   PageHeader,
@@ -150,7 +149,6 @@ export const DialogLayoutDemo = ({ router }) => (
         alert('Success!'); // eslint-disable-line no-alert
         router.push('/layouts/dialogLayout');
       }}
-      logo={<CJLogo />}
     >
       <DemoForm />
     </DialogLayout>
@@ -172,7 +170,6 @@ export const SubmittingDialogLayoutDemo = ({ router }) => (
         alert('Success!'); // eslint-disable-line no-alert
         router.push('/layouts/dialogLayout');
       }}
-      logo={<CJLogo />}
     >
       <DemoForm />
     </DialogLayout>
@@ -186,7 +183,6 @@ export const NoButtonDialogLayoutDemo = ({ router }) => (
     <DialogLayout
       title={'Create Program Term'}
       onCancel={() => router.push('/layouts/dialogLayout')}
-      logo={<CJLogo />}
     >
       <Panel>
         <Header>Panel Header</Header>
@@ -209,7 +205,6 @@ export const WideDialogLayoutDemo = ({ router }) => (
         alert('Success!'); // eslint-disable-line no-alert
         router.push('/layouts/dialogLayout');
       }}
-      logo={<CJLogo />}
       contentSize={'wide'}
     >
       <DemoForm />