+++ packages/visual-stack/src/layouts/DialogLayout/index.js
@@ -5,6 +5,7 @@ import cn from 'classnames';
 
 import './index.css';
 import Spinner from '../../components/Spinner';
+import { TopNav } from '../../components/TopNav'
 
 function getSubmitButtonText(submitButtonText, showSubmitButtonSpinner) {
   if (showSubmitButtonSpinner) {
@@ -25,47 +26,23 @@ export const DialogLayout = ({
   disableSubmit,
   showSubmitButtonSpinner,
   contentSize,
-  logo,
   onCancel,
   onSubmit,
   children,
   ...restProps
 }) => (
   <div className={cn(`vs-dialog-layout`, className)} {...restProps}>
-    <div className="vs-dialog-layout-header vs-dialog-layout-page-title">
-      <div className="vs-dialog-layout-logo-title-container">
-        <div className="vs-dialog-layout-logo-container">
-          <span className="vs-cj-logo">{logo}</span>
-        </div>
-
-        <h1 className="vs-dialog-layout-title">{title}</h1>
-      </div>
-
-      <div className="vs-dialog-layout-button-bar">
-        {cancelButtonText && (
-          <Button id="vs-dialog-layout-cancel" type="text" onClick={onCancel}>
-            {cancelButtonText}
-          </Button>
-        )}
-
-        {submitButtonText && (
-          <Button
-            id="vs-dialog-layout-submit"
-            type="solid-primary"
-            disabled={disableSubmit === true}
-            onClick={onSubmit}
-          >
-            {getSubmitButtonText(submitButtonText, showSubmitButtonSpinner)}
-          </Button>
-        )}
-        {!submitButtonText && !cancelButtonText && (
-          <CloseIcon
-            className="vs-dialog-layout-icon-close"
-            onClick={onCancel}
-          />
-        )}
-      </div>
-    </div>
+    <TopNav
+      title={title}
+      actionChildren=<TopNavControls
+        onSubmit={onSubmit}
+        onCancel={onCancel}
+        submitButtonText={submitButtonText}
+        cancelButtonText={cancelButtonText}
+        disableSubmit={disableSubmit}
+        showSubmitButtonSpinner={showSubmitButtonSpinner}
+      />
+    />
     <div
       className={cn(
         'vs-dialog-layout-content',
@@ -76,5 +53,39 @@ export const DialogLayout = ({
     </div>
   </div>
 );
+
+const TopNavControls = ({
+                          onSubmit,
+                          onCancel,
+                          submitButtonText,
+                          cancelButtonText,
+                          disableSubmit,
+                          showSubmitButtonSpinner,
+                        }) => (
+  <div className="vs-dialog-layout-button-bar">
+    {cancelButtonText && (
+      <Button id="vs-dialog-layout-cancel" type="text" onClick={onCancel}>
+        {cancelButtonText}
+      </Button>
+    )}
+
+    {submitButtonText && (
+      <Button
+        id="vs-dialog-layout-submit"
+        type="solid-primary"
+        disabled={disableSubmit === true}
+        onClick={onSubmit}
+      >
+        {getSubmitButtonText(submitButtonText, showSubmitButtonSpinner)}
+      </Button>
+    )}
+    {!submitButtonText && !cancelButtonText && (
+      <CloseIcon
+        className="vs-dialog-layout-icon-close"
+        onClick={onCancel}
+      />
+    )}
+  </div>
+)
 
 export default DialogLayout;